api/requirements.txt
ephem==4.1.5
google-generativeai==0.8.3
requests==2.32.3
